requirements.txt
inference-sdk
opencv-python==4.10.0.84
numpy<2.0.0
requests>=2.32.0
spotipy==2.23.0
psutil==5.9.5
screen-brightness-control==0.21.0
WMI==1.5.1
pywin32==306
python-dotenv==1.0.0
